+++ src/app/data_harvester/disks/unix.rs
@@ -1,41 +1,121 @@
-//! Implementation based on [heim's](https://github.com/heim-rs/heim)
-//! Unix disk usage.
+//! Disk stats for Unix-like systems that aren't supported through other means.
 
-use std::path::Path;
+use crate::app::Filter;
+use crate::data_harvester::disks::{DiskHarvest, IoData, IoHarvest};
 
-use crate::app::{
-    data_harvester::disks::{DiskHarvest, IoHarvest},
-    filter::Filter,
-};
+pub fn get_io_usage() -> anyhow::Result<IoHarvest> {
+    // if !actually_get {
+    //     return Ok(None);
+    // }
 
-mod file_systems;
-pub(crate) use file_systems::FileSystem;
+    // let mut io_hash: std::collections::HashMap<String, Option<IoData>> =
+    //     std::collections::HashMap::new();
 
-cfg_if::cfg_if! {
-    if #[cfg(target_os = "linux")] {
-        use super::linux::partitions;
-    } else {
-        mod partition;
-        mod mounts;
+    // let counter_stream = heim::disk::io_counters().await?;
 
-        use partition::*;
-    }
-}
+    // while let Some(io) = counter_stream.next().await {
+    //     if let Ok(io) = io {
+    //         let mount_point = io.device_name().to_str().unwrap_or("Mount Unavailable");
 
-pub(crate) struct Usage {}
+    //         io_hash.insert(
+    //             mount_point.to_string(),
+    //             Some(IoData {
+    //                 read_bytes: io.read_bytes().get::<heim::units::information::byte>(),
+    //                 write_bytes: io.write_bytes().get::<heim::units::information::byte>(),
+    //             }),
+    //         );
+    //     }
+    // }
 
-fn disk_usage(path: &Path) -> anyhow::Result<Usage> {
+    // Ok(Some(io_hash))
+
     todo!()
 }
 
 pub fn get_disk_usage(
     disk_filter: &Option<Filter>, mount_filter: &Option<Filter>,
 ) -> anyhow::Result<Vec<DiskHarvest>> {
-    let partitions = partitions()?;
+    todo!()
+    // let mut vec_disks: Vec<DiskHarvest> = Vec::new();
+    // let partitions_stream = heim::disk::partitions_physical().await?;
 
-    Ok(vec![])
-}
+    // while let Some(part) = partitions_stream.next().await {
+    //     if let Ok(partition) = part {
+    //         let name = get_device_name(&partition);
 
-pub fn get_io_usage() -> anyhow::Result<IoHarvest> {
-    Ok(IoHarvest::default())
+    //         let mount_point = (partition
+    //             .mount_point()
+    //             .to_str()
+    //             .unwrap_or("Name unavailable"))
+    //         .to_string();
+
+    //         // Precedence ordering in the case where name and mount filters disagree, "allow" takes precedence over "deny".
+    //         //
+    //         // For implementation, we do this as follows:
+    //         // 1. Is the entry allowed through any filter? That is, does it match an entry in a filter where `is_list_ignored` is `false`? If so, we always keep this entry.
+    //         // 2. Is the entry denied through any filter? That is, does it match an entry in a filter where `is_list_ignored` is `true`? If so, we always deny this entry.
+    //         // 3. Anything else is allowed.
+
+    //         let filter_check_map = [(disk_filter, &name), (mount_filter, &mount_point)];
+
+    //         // This represents case 1. That is, if there is a match in an allowing list - if there is, then
+    //         // immediately allow it!
+    //         let matches_allow_list = filter_check_map.iter().any(|(filter, text)| {
+    //             if let Some(filter) = filter {
+    //                 if !filter.is_list_ignored {
+    //                     for r in &filter.list {
+    //                         if r.is_match(text) {
+    //                             return true;
+    //                         }
+    //                     }
+    //                 }
+    //             }
+    //             false
+    //         });
+
+    //         let to_keep = if matches_allow_list {
+    //             true
+    //         } else {
+    //             // If it doesn't match an allow list, then check if it is denied.
+    //             // That is, if it matches in a reject filter, then reject.  Otherwise, we always keep it.
+    //             !filter_check_map.iter().any(|(filter, text)| {
+    //                 if let Some(filter) = filter {
+    //                     if filter.is_list_ignored {
+    //                         for r in &filter.list {
+    //                             if r.is_match(text) {
+    //                                 return true;
+    //                             }
+    //                         }
+    //                     }
+    //                 }
+    //                 false
+    //             })
+    //         };
+
+    //         if to_keep {
+    //             // The usage line can fail in some cases (for example, if you use Void Linux + LUKS,
+    //             // see https://github.com/ClementTsang/bottom/issues/419 for details).  As such, check
+    //             // it like this instead.
+    //             if let Ok(usage) = heim::disk::usage(partition.mount_point()).await {
+    //                 vec_disks.push(DiskHarvest {
+    //                     free_space: Some(usage.free().get::<heim::units::information::byte>()),
+    //                     used_space: Some(usage.used().get::<heim::units::information::byte>()),
+    //                     total_space: Some(usage.total().get::<heim::units::information::byte>()),
+    //                     mount_point,
+    //                     name,
+    //                 });
+    //             } else {
+    //                 vec_disks.push(DiskHarvest {
+    //                     free_space: None,
+    //                     used_space: None,
+    //                     total_space: None,
+    //                     mount_point,
+    //                     name,
+    //                 });
+    //             }
+    //         }
+    //     }
+    // }
+
+    // Ok(Some(vec_disks))
 }
